backend/requirements.txt

torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
opencv-python>=4.8.0
albumentations>=1.3.1
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
timm>=0.9.0
